refactor: move NavLink component to nav directory
